+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom"
 import { Provider } from "react-redux"
 import { PersistGate } from "redux-persist/lib/integration/react"
 import { store, persistor } from "./redux/store"
-import App from "./App.jsx"
+import App from "./App"
 import "css/global.css"
 
 ReactDOM.render(